% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch_and_larch_and_picea_data.R
\docType{data}
\name{picea}
\alias{picea}
\title{Qinghai Spruce Dataset}
\format{
A data frame with 402 rows and 24 variables:
\describe{
  \item{LIDAR-X}{Projected X-coordinate of the LIDAR-detected tree, in meters.}
  \item{LIDAR-Y}{Projected Y-coordinate of the LIDAR-detected tree, in meters.}
  \item{LH}{Tree height derived from LIDAR data, in meters.}
  \item{LHCB}{Crown base height derived from LIDAR data, in meters.}
  \item{LCW1}{East-west crown diameter derived from LIDAR data, in meters.}
  \item{LCW2}{South-north crown diameter derived from LIDAR data, in meters.}
  \item{LCW}{Mean LIDAR-derived crown diameter, calculated as the mean of \code{LCW1} and \code{LCW2}, in meters.}
  \item{CPA}{Crown projection area, in square meters.}
  \item{X}{Projected X-coordinate of the ground-measured tree, in meters.}
  \item{Y}{Projected Y-coordinate of the ground-measured tree, in meters.}
  \item{Z}{Elevation of the tree base, in meters above sea level.}
  \item{PLOT}{Character sample plot name, from "yd01" to "yd16".}
  \item{PLOT1}{Numeric sample plot identifier, from 1 to 16.}
  \item{OBS}{Tree observation identifier.}
  \item{D0}{Ground-measured diameter at breast height (DBH), in centimeters.}
  \item{H0}{Ground-measured height of the tree, in meters.}
  \item{HCB0}{Crown base height (branch-free height) measured on the ground, in meters.}
  \item{CW01}{Ground-measured east-west crown diameter, in meters.}
  \item{CW02}{Ground-measured south-north crown diameter, in meters.}
  \item{CW}{Mean ground-measured crown diameter, calculated as the mean of \code{CW01} and \code{CW02}, in meters.}
  \item{STEM}{Estimated biomass of the tree stem, in kilograms.}
  \item{BRANCH}{Estimated biomass of the tree branches, in kilograms.}
  \item{FOLIAGE}{Estimated biomass of the tree foliage, in kilograms.}
  \item{FRUIT}{Estimated biomass of the tree fruit, in kilograms.}
}
}
\usage{
picea
}
\description{
A dataset containing information about Qinghai spruce trees at the Xishui forest farm of Su'nan Yuguzu Autonomous County, China, in Gansu Qilian Mountain National Nature Reserve.
}
\keyword{datasets}
