% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1.R
\docType{data}
\name{plot_1}
\alias{plot_1}
\title{1st period sample plot survey data}
\format{
`plot_1`
A data frame with 62 rows and 23 columns:
\describe{
  \item{plot_id}{Plot ID}
  \item{standing_stock}{Standing stock}
  \item{forest_cutting_stock}{Forest cutting stock}
  \item{crown_density}{Crown density}
  \item{disaster_level}{Disaster level}
  \item{origin}{origin}
  \item{dominant_tree_species}{Dominant tree species}
  \item{age_group}{Age group}
  \item{naturalness}{Naturalness}
  \item{land_type}{Land type}
  ...
}
}
\usage{
plot_1
}
\description{
The 1st period sample plot survey data (e.g. 2005)
}
\keyword{datasets}
