% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2.R
\docType{data}
\name{plot_2}
\alias{plot_2}
\title{2nd period sample plot survey data}
\format{
`plot_2`
A data frame with 100 rows and 5 columns:
\describe{
  \item{plot_id}{Plot ID}
  \item{standing_stock}{Standing stock}
  \item{forest_cutting_stock}{Forest cutting stock}
  \item{crown_density}{Crown density}
  \item{disaster_level}{Disaster level}
  \item{origin}{origin}
  \item{dominant_tree_species}{Dominant tree species}
  \item{age_group}{Age group}
  \item{naturalness}{Naturalness}
  \item{land_type}{Land type}
  ...
}
}
\usage{
plot_2
}
\description{
The 2nd period sample plot survey data (e.g. 2010)
}
\keyword{datasets}
