% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_1.R
\docType{data}
\name{tree_1}
\alias{tree_1}
\title{1st period trees survey data}
\format{
`tree_1`
A data frame with 1634 rows and 5 columns:
\describe{
  \item{plot_id}{Plot ID}
  \item{inspection_type}{Inspection type}
  \item{tree_species_code}{Tree species code}
  ...
}
}
\usage{
tree_1
}
\description{
The 1st period trees survey data (e.g. 2005)
}
\keyword{datasets}
