% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_2.R
\docType{data}
\name{tree_2}
\alias{tree_2}
\title{2nd period trees survey data}
\format{
`tree_2`
A data frame with 4778 rows and 5 columns:
\describe{
  \item{plot_id}{Plot ID}
  \item{inspection_type}{Inspection type}
  \item{tree_species_code}{Tree species code}
  ...
}
}
\usage{
tree_2
}
\description{
The 2nd period trees survey data (e.g. 2010)
}
\keyword{datasets}
