% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_3.R
\docType{data}
\name{tree_3}
\alias{tree_3}
\title{3rd period trees survey data}
\format{
`tree_3`
A data frame with 4528 rows and 5 columns:
\describe{
  \item{plot_id}{Plot ID}
  \item{inspection_type}{Inspection type}
  \item{tree_species_code}{Tree species code}
  ...
}
}
\usage{
tree_3
}
\description{
The 3rd period trees survey data (e.g. 2015)
}
\keyword{datasets}
