% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestData.R
\docType{data}
\name{forestData}
\alias{forestData}
\title{Subzone 6 Oak broadleaf mixed data}
\format{
`forestData`
A data frame with 1,326 rows and 79 columns:
\describe{
  \item{ID}{Unique identifier for each tree}
  \item{AGE}{Stand age of the tree}
  \item{H}{Height of the tree}
  \item{BA}{Basal area of the tree}
  \item{Bio}{Biomass of the tree}
  \item{S}{Forest density index}
  \item{code}{Codes for forest types}
  ...
}
}
\usage{
forestData
}
\description{
Subzone 6 Oak broadleaf mixed data from fieldwork
}
\keyword{datasets}
