% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestData.R
\docType{data}
\name{forestData}
\alias{forestData}
\title{Mixed birch-broadleaf forest data}
\format{
`forestData`
A data frame with 320 rows and 16 columns:
\describe{
  \item{ID}{Plot ID}
  \item{AGE}{The average age of the stand}
  \item{H}{Stand height}
  \item{BA}{Stand basal area}
  \item{Bio}{Stand biomass}
  \item{S}{Stand density index}
  \item{code}{Forest type code of plot}
  ...
}
}
\usage{
forestData
}
\description{
Mixed birch-broadleaf forest data
}
\keyword{datasets}
