% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch_and_larch_and_picea_data.R
\docType{data}
\name{larch}
\alias{larch}
\title{Larch Dataset}
\format{
A data frame with 4538 rows and 18 variables:
\describe{
  \item{PLOT}{Sample plot identifier, indicating the plot number.}
  \item{TREE}{Tree species, which is "larch" in this dataset.}
  \item{D}{Diameter at breast height (DBH) of the tree, in centimeters.}
  \item{H}{Height of the tree, in meters.}
  \item{CBH}{Crown base height (branch-free height), in meters.}
  \item{LIFE}{Health status of the tree (e.g., "alive", "healthy", "unknown").}
  \item{CW_E}{Crown width in the east direction, in meters.}
  \item{CW_S}{Crown width in the south direction, in meters.}
  \item{CW_W}{Crown width in the west direction, in meters.}
  \item{CW_N}{Crown width in the north direction, in meters.}
  \item{X}{X-coordinate for individual tree location within the plot.}
  \item{Y}{Y-coordinate for individual tree location within the plot.}
  \item{SD}{Stand density, representing the density of trees within the plot.}
  \item{CLR}{Crown length ratio, which indicates the ratio of crown length to tree height.}
  \item{CW}{Overall crown width, in meters.}
  \item{CW_EW}{East-West crown width, in meters.}
  \item{CW_SN}{South-North crown width, in meters.}
  \item{AGE.GROUP}{Age group of the tree, which can be one of the following: "成熟林", "过熟林", "近熟林", "幼龄林", "中龄林".}
}
}
\usage{
larch
}
\description{
A dataset containing information about larch trees in the Winter Olympics core area.
}
\keyword{datasets}
