% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birch_and_larch_and_picea_data.R
\docType{data}
\name{picea}
\alias{picea}
\title{Qinghai Spruce Dataset}
\format{
A data frame with 402 rows and 24 variables:
\describe{
  \item{LIDAR-X}{LIDAR X-coordinate of the tree, representing spatial location data.}
  \item{LIDAR-Y}{LIDAR Y-coordinate of the tree, representing spatial location data.}
  \item{LH}{Tree height derived from LIDAR data, in meters.}
  \item{LHCB}{Crown base height derived from LIDAR data, in meters.}
  \item{LCW1}{Crown width measurement 1 derived from LIDAR data, in meters.}
  \item{LCW2}{Crown width measurement 2 derived from LIDAR data, in meters.}
  \item{LCW}{Average crown width derived from LIDAR data, in meters.}
  \item{CPA}{Crown projection area, representing the horizontal area of the crown, in square meters.}
  \item{X}{Ground X-coordinate of individual tree location within the plot.}
  \item{Y}{Ground Y-coordinate of individual tree location within the plot.}
  \item{Z}{Elevation of the tree base, in meters above sea level.}
  \item{PLOT}{Sample plot identifier, indicating the plot number.}
  \item{PLOT1}{Plot name or ID within the study area.}
  \item{OBS}{Observation number for the specific tree within the plot.}
  \item{D0}{Diameter at breast height (DBH) of the tree, in meters.}
  \item{H0}{Ground-measured height of the tree, in meters.}
  \item{HCB0}{Crown base height (branch-free height) measured on the ground, in meters.}
  \item{CW01}{Crown width measurement 1 measured on the ground, in meters.}
  \item{CW02}{Crown width measurement 2 measured on the ground, in meters.}
  \item{CW}{Average crown width measured on the ground, in meters.}
  \item{STEM}{Estimated biomass of the tree stem, in kilograms.}
  \item{BRANCH}{Estimated biomass of the tree branches, in kilograms.}
  \item{FOLIAGE}{Estimated biomass of the tree foliage, in kilograms.}
  \item{FRUIT}{Estimated biomass of the tree fruit, in kilograms.}
}
}
\usage{
picea
}
\description{
A dataset containing information about Qinghai spruce trees at the Xishui forest farm of Su’nan Yuguzu Autonomous County, China, in Gansu Qilian Mountain National Nature Reserve.
}
\keyword{datasets}
