% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.forestData}
\alias{plot.forestData}
\title{ForestData Plot}
\usage{
\method{plot}{forestData}(
  x,
  model.type = "H",
  plot.type = "Curve",
  xlab = NA,
  ylab = NA,
  legend.lab = "Site class",
  title = "Oak broadleaf mixed",
  ...
)
}
\arguments{
\item{x}{A data of forestData class.}

\item{model.type}{A character in one of the three characters H,BA,Bio.This means that the figure should use the model you choose to plot.}

\item{plot.type}{A character in one of the four characters "Curve","residuals","Scatter_Curve","Scatter".This means that the picture is drawn of the type you have chosen}

\item{xlab}{The title for the x axis.}

\item{ylab}{The title for the y axis.}

\item{legend.lab}{The title for the legends.}

\item{title}{The text for the title.}

\item{...}{Additional arguments affecting the figure plotted.}
}
\description{
Plot example figures about the forestData.
}
