% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.forestData}
\alias{summary.forestData}
\title{ForestData Summary}
\usage{
\method{summary}{forestData}(object, ...)
}
\arguments{
\item{object}{Data of a forestData class(after class.plot).}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Gives summary information about the forestData.
}
\details{
The summary includes the summary of raw data, the model, the model parameters, potential productivity and real productivity in forestData(if available)
}
